# api.yaml
openapi: 3.0.3
info:
  title: Viking Demo API
  version: 1.0.0
  description: REST API для работы с таблицей викингов
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
paths:
  /api/vikings:
    get:
      tags: [Vikings]
      summary: Получить список всех созданных викингов
      operationId: getAllVikings
      responses:
        '200':
          description: Список викингов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viking'
    post:
      tags: [Vikings]
      summary: Добавить конкретного викинга
      operationId: createViking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VikingRequest'
      responses:
        '201':
          description: Викинг успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
  /api/vikings/{id}:
    get:
      tags: [Vikings]
      summary: Получить викинга по id
      operationId: getVikingById
      parameters:
        - $ref: '#/components/parameters/VikingId'
      responses:
        '200':
          description: Викинг найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
        '404':
          description: Викинг не найден
    put:
      tags: [Vikings]
      summary: Полностью перезаписать параметры конкретного викинга
      operationId: updateViking
      parameters:
        - $ref: '#/components/parameters/VikingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VikingRequest'
      responses:
        '200':
          description: Викинг успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
        '404':
          description: Викинг не найден
    delete:
      tags: [Vikings]
      summary: Удалить конкретного викинга из таблицы
      operationId: deleteViking
      parameters:
        - $ref: '#/components/parameters/VikingId'
      responses:
        '204':
          description: Викинг успешно удалён
        '404':
          description: Викинг не найден
components:
  parameters:
    VikingId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int32
      example: 1
  schemas:
    Viking:
      allOf:
        - $ref: '#/components/schemas/VikingRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1
    VikingRequest:
      type: object
      required: [name, age, heightCm, hairColor, beardStyle, equipment]
      properties:
        name:
          type: string
          example: Ragnar
        age:
          type: integer
          format: int32
          example: 34
        heightCm:
          type: integer
          format: int32
          example: 182
        hairColor:
          $ref: '#/components/schemas/HairColor'
        beardStyle:
          $ref: '#/components/schemas/BeardStyle'
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentItem'
    EquipmentItem:
      type: object
      required: [name, quality]
      properties:
        name:
          type: string
          example: Axe
        quality:
          type: string
          example: Rare
    HairColor:
      type: string
      enum: [Blond, Red, Brown, Black, Gray]
      example: Blond
    BeardStyle:
      type: string
      enum: [SHORT, LONG, BRAIDED, FORKED, CLEAN_SHAVEN]
      example: BRAIDED